+++ Controllers/productsController.js
@@ -2,19 +2,11 @@ import express from 'express';
 import database from '../database.js';
 import Joi from 'joi';
 
-let itemSchema = Joi.object({
-    brand: Joi.string().min(1).required(),
-    model: Joi.string().min(1).required(),
-    amount: Joi.number().required(),
-    price: Joi.string().min(1).required() 
-}); 
-
 export async function productsPOST(req, res){
     let item = req.body;
-    let { error } = itemSchema.validate(item);
     const { authorization } = req.headers;
 	const token = authorization?.replace('Bearer', '').trim();
-    if(error === undefined && token){
+    if(token){
         try{
             let access = await database.collection('session').find({token: token}).toArray();
             if(access.length !== 0){
